refactor: category use case
- Fix model calls to entity
- Configure the entities in typeorm
- Reference the category table to the category entity
- Removes the instantiation of a local repository to use the database in the application, using typeorm.
- Put the use cases of creating a category in a function, so that is called only when registering a category.
- This allows for proper connection to the database and corrects the error where the repository was instantiated even before the database was up.
