+++ ormconfig.json
@@ -6,6 +6,7 @@
   "password": "ignite",
   "database": "rentx",
   "migrations": ["./src/database/migrations/*.ts"],
+  "entities": ["./src/modules/**/entities/*.ts"],
   "cli": {
     "migrationsDir": "./src/database/migrations"
   }
