+++ src/modules/cars/entities/Category.ts
@@ -1,7 +1,7 @@
 import { Column, CreateDateColumn, Entity, PrimaryColumn } from 'typeorm';
 import { v4 as uuidV4 } from 'uuid';
 
-@Entity()
+@Entity('categories')
 class Category {
   @PrimaryColumn()
   id?: string;
