+++ src/modules/cars/useCases/createCategory/CreateCategoryController.ts
@@ -5,11 +5,11 @@ import { CreateCategoryUseCase } from './CreateCategoryUseCase';
 class CreateCategoryController {
   constructor(private createCategoryUseCase: CreateCategoryUseCase) {}
 
-  handle(request: Request, response: Response): Response {
+  async handle(request: Request, response: Response): Promise<Response> {
     try {
       const { name, description } = request.body;
 
-      this.createCategoryUseCase.execute({ name, description });
+      await this.createCategoryUseCase.execute({ name, description });
 
       return response.status(201).send();
     } catch (err) {
