+++ src/server.ts
@@ -1,10 +1,10 @@
 import express from 'express';
 import swaggerUi from 'swagger-ui-express';
 
+import './database';
+
 import { router } from './routes';
 import swaggerFile from './swagger.json';
-
-import './database';
 
 const PORT = 3333;
